Add functions for updating and installing packages, cleaning workspace, cloning repositories, copying modules and config, generating images, compressing and copying files, and generating all images

+++ runloop.sh
@@ -1,4 +1,39 @@
+# Function to update and install required packages
+updateAndInstallPackages() {
+    sudo apt update
+    sudo apt upgrade --yes
+    sudo apt install --yes aria2 coreutils jq p7zip-full qemu-user-static zip dialog xz xz-utils nodejs
+    sudo npm install -g n
+}
 
+# Function to clean the workspace
+cleanWorkspace() {
+    cd ~/GAB-OS-MOD/
+    ./clean.sh
+}
+
+# Function to clone the GAB-OS repository
+cloneGABOS() {
+    cd ~/
+    git clone https://github.com/gab-3d/GAB-OS.git
+}
+
+# Function to copy modules and config
+copyModulesAndConfig() {
+    cp -r ~/GAB-OS-MOD/modules/* ~/GAB-OS/src/modules/
+    cp -r ~/GAB-OS-MOD/config/* ~/GAB-OS/config/raspberry/
+}
+
+# Function to clone the CustomPiOS repository
+cloneCustomPiOS() {
+    cd ~
+    git clone https://github.com/guysoft/CustomPiOS.git
+    cd CustomPiOS
+    git branch stable 63da54b86ab566c558e9084568d326413f6585d8
+    git switch stable
+}
+
+# Function to generate an image
 generateImage() {
     local version=$1
     cd ~/GAB-OS/
@@ -36,16 +71,14 @@ generateImage() {
     echo "DOWNLOAD_URL_IMAGE=${DOWNLOAD_URL_IMAGE}" >> $GITHUB_OUTPUT
     echo "MODULES=${MODULES}" >> $GITHUB_OUTPUT
 
-
     cd ./src/image
     FILENAME=$(basename $DOWNLOAD_URL_CHECKSUM )
-    wget -O $FILENAME $DOWNLOAD_URL_CHECKSUM }
+    wget -O $FILENAME $DOWNLOAD_URL_CHECKSUM
     FILE_CONTENT=$(head -n 1 $FILENAME)
     CHECKSUM=$(echo $FILE_CONTENT | cut -d' ' -f1)
 
     echo "CHECKSUM=$CHECKSUM" >> $GITHUB_OUTPUT
     echo "FILENAME=$FILENAME" >> $GITHUB_OUTPUT
-
 
     aria2c -d . --seed-time=0 $DOWNLOAD_URL_IMAGE
 
@@ -54,53 +87,58 @@ generateImage() {
     cd ./src && ../../CustomPiOS/src/update-custompios-paths
     sudo modprobe loop && sudo bash -x ./build_dist
 
+    if [ $? -ne 0 ]; then
+        echo "Build failed"
+        exit 1
+    fi
 
-    #create a filename tat start wirh current date and time add gab-os-mod to the end 
+    compressAndCopyFiles
+
+}
+# Function to compress and copy image and log files
+compressAndCopyFiles() {
     TARGGET_FILENAME=$(date +"%Y-%m-%d-%H-%M-gab-os-")$SBC
 
     cd ~/GAB-OS/src/workspace/
-    CPU_COUNT=12
+    CPU_COUNT=$(grep -c ^processor /proc/cpuinfo)
     echo -e "\e[32mUsing ${CPU_COUNT} Cores for compression...\e[0m"
     sudo xz -efkvz9T"${CPU_COUNT}" '2023-05-03-raspios-bullseye-arm64-lite.img' || true
 
-
     cp ~/GAB-OS/src/workspace/2023-05-03-raspios-bullseye-arm64-lite.img.xz ~/GAB-OS-MOD/$TARGGET_FILENAME.img.xz
-    cp ~/GAB-OS/src/build.log ~/GAB-OS-MOD/$TARGGET_FILENAME.log
 
-        # Rest of the code goes here
+    cp ~/GAB-OS/src/build.log ~/GAB-OS-MOD/$TARGGET_FILENAME.log
+}
+# Function to get the images to generate
+getImagesToGenerate() {
+    # Ask for which image to generate with dialog
+    dialog --checklist "Select the images to generate" 0 0 0 \
+    "raspberry/rpi64-base" "Base image" on \
+    "raspberry/rpi64-ks-voron" "Voron image" on \
+    "raspberry/rpi64-ks-mk3" "MK3 image" on 3>&1 1>&2 2>&3
 }
 
-sudo apt update; sudo apt install --yes aria2 coreutils jq p7zip-full qemu-user-static zip dialog
+# Function to generate all images
+generateAllImages() {
+    imageToGenerate=($(getImagesToGenerate))
 
-cd ~/GAB-OS-MOD/
-./clean.sh
+    for image in "${imageToGenerate[@]}"
+    do
+        generateImage "$image"
+        if [ $? -ne 0 ]; then
+            echo "Build failed"
+            exit 1
+        fi
+        sudo rm -rf ~/GAB-OS/src/workspace/*
 
 
-cd ~/
 
-git clone https://github.com/gab-3d/GAB-OS.git
+    done
+}
 
-cp -r ~/GAB-OS-MOD/modules/* ~/GAB-OS/src/modules/
-cp -r ~/GAB-OS-MOD/config/* ~/GAB-OS/config/raspberry/
-
-
-cd ~
-git clone https://github.com/guysoft/CustomPiOS.git
-
-cd CustomPiOS
-git branch stable 63da54b86ab566c558e9084568d326413f6585d8
-git switch stable
-
-imageToGenerate=("raspberry/rpi64-base" "raspberry/rpi64-ks-voron" "raspberry/rpi64-ks-mk3")
-imageToGenerate=("raspberry/rpi64-ks-mk3")
-
-for image in "${imageToGenerate[@]}"
-do
-    generateImage "$image"
-    sudo rm -rf ~/GAB-OS/src/workspace/*
-done
-# sudo umount /home/gab-os/GAB-OS/src/workspace/mount/dev/pts
-# sudo umount /home/gab-os/GAB-OS/src/workspace/mount/boot
-# sudo umount /home/gab-os/GAB-OS/src/workspace/mount
-#generateImage "raspberry/rpi64-ks-voron"
-
+# Call all functions
+updateAndInstallPackages
+cleanWorkspace
+cloneGABOS
+copyModulesAndConfig
+cloneCustomPiOS
+generateAllImages
